Tranfered guild endpoint
