+++ clamor/models/emoji.py
@@ -1,0 +1,5 @@
+from .base import Base
+
+
+class Emoji(Base):
+    pass
