+++ clamor/models/guild.py
@@ -1,0 +1,25 @@
+from .base import Base
+
+
+class Ban(Base):
+    pass
+
+
+class Integration(Base):
+    pass
+
+
+class Role(Base):
+    pass
+
+
+class Member(Base):
+    pass
+
+
+class Guild(Base):
+    pass
+
+
+class GuildEmbed(Base):
+    pass
