+++ clamor/models/voice.py
@@ -1,0 +1,5 @@
+from .base import Base
+
+
+class VoiceRegion(Base):
+    pass
